refactor: Split xcp implementation in eth-transport and xcp
